MAINT: Overwrite old pr reminder with new code
